requirements.txt
Django==5.2
mistralai==1.7.0
ollama==0.4.8
python-dotenv==1.1.0
requests==2.32.3
requests
SQLAlchemy==2.0.40
sqlparse==0.5.3
django-extensions # Extensions for Django
django-formtools # Form handling library for Django
tenacity==9.1.2  # Retry library for handling transient errors - useful for network calls (api)
sqlparse # SQL parsing library
tzdata # Time zone data for PostgreSQL
djangorestframework # REST framework for building APIs
gunicorn # WSGI server for production


